refactor: mv user structs to separate file
