+++ include/kitepp/responses.hpp
@@ -48,198 +48,6 @@ namespace rj = rapidjson;
 namespace kc = kiteconnect;
 namespace rju = kc::rjutils;
 
-/// userProfile represents a user's personal and financial profile.
-struct userProfile {
-
-    userProfile() = default;
-
-    explicit userProfile(const rj::Value::Object& val) { parse(val); };
-
-    void parse(const rj::Value::Object& val) {
-
-        rju::_getIfExists(val, userID, "user_id");
-        rju::_getIfExists(val, userName, "user_name");
-        rju::_getIfExists(val, userShortName, "user_shortname");
-        rju::_getIfExists(val, avatarURL, "avatar_url");
-        rju::_getIfExists(val, userType, "user_type");
-        rju::_getIfExists(val, email, "email");
-        rju::_getIfExists(val, broker, "broker");
-        rju::_getIfExists(val, products, "products");
-        rju::_getIfExists(val, orderTypes, "order_types");
-        rju::_getIfExists(val, exchanges, "exchanges");
-
-        rj::Value metaVal(rj::kObjectType);
-        rju::_getIfExists(val, metaVal, "meta", rju::_RJValueType::Object);
-        meta.parse(metaVal.GetObject());
-    };
-
-    string userID;
-    string userName;
-    string userShortName;
-    string avatarURL;
-    string userType;
-    string email;
-    string broker;
-    std::vector<string> products;
-    std::vector<string> orderTypes;
-    std::vector<string> exchanges;
-    struct Meta {
-
-        Meta() = default;
-
-        explicit Meta(const rj::Value::Object& val) { parse(val); };
-
-        void parse(const rj::Value::Object& val) { rju::_getIfExists(val, dematConsent, "demat_consent"); }
-
-        string dematConsent;
-    } meta;
-};
-
-/// tokens received after successfull authentication
-struct userTokens {
-
-    userTokens() = default;
-
-    explicit userTokens(const rj::Value::Object& val) { parse(val); };
-
-    void parse(const rj::Value::Object& val) {
-
-        rju::_getIfExists(val, userID, "user_id");
-        rju::_getIfExists(val, accessToken, "access_token");
-        rju::_getIfExists(val, refreshToken, "refresh_token");
-    };
-
-    string userID;
-    string accessToken;
-    string refreshToken;
-};
-
-/// userSession represents the response after a successful authentication.
-struct userSession {
-
-    userSession() = default;
-
-    explicit userSession(const rj::Value::Object& val) { parse(val); };
-
-    void parse(const rj::Value::Object& val) {
-
-        profile.parse(val);
-        tokens.parse(val);
-        rju::_getIfExists(val, apiKey, "api_key");
-        rju::_getIfExists(val, publicToken, "public_token");
-        rju::_getIfExists(val, loginTime, "login_time");
-    };
-
-    userProfile profile;
-    userTokens tokens;
-
-    string apiKey;
-    string publicToken;
-    string loginTime;
-};
-
-/// availableMargins represents the available margins from the margins response for a single segment.
-struct availableMargins {
-
-    availableMargins() = default;
-
-    explicit availableMargins(const rj::Value::Object& val) { parse(val); };
-
-    void parse(const rj::Value::Object& val) {
-
-        double debugval = val["cash"].GetDouble();
-
-        rju::_getIfExists(val, adHocMargin, "adhoc_margin");
-        rju::_getIfExists(val, cash, "cash");
-        rju::_getIfExists(val, collateral, "collateral");
-        rju::_getIfExists(val, intradayPayin, "intraday_payin");
-    };
-
-    double adHocMargin = 0.0;
-    double cash = 0.0;
-    double collateral = 0.0;
-    double intradayPayin = 0.0;
-};
-
-/// usedMargins represents the used margins from the margins response for a single segment.
-struct usedMargins {
-
-    usedMargins() = default;
-
-    explicit usedMargins(const rj::Value::Object& val) { parse(val); };
-
-    void parse(const rj::Value::Object& val) {
-        rju::_getIfExists(val, debits, "debits");
-        rju::_getIfExists(val, exposure, "exposure");
-        rju::_getIfExists(val, M2MRealised, "m2m_realised");
-        rju::_getIfExists(val, M2MUnrealised, "m2m_unrealised");
-        rju::_getIfExists(val, optionPremium, "option_premium");
-        rju::_getIfExists(val, payout, "payout");
-        rju::_getIfExists(val, span, "span");
-        rju::_getIfExists(val, holdingSales, "holding_sales");
-        rju::_getIfExists(val, turnover, "turnover");
-    };
-
-    double debits = 0.0;
-    double exposure = 0.0;
-    double M2MRealised = 0.0;
-    double M2MUnrealised = 0.0;
-    double optionPremium = 0.0;
-    double payout = 0.0;
-    double span = 0.0;
-    double holdingSales = 0.0;
-    double turnover = 0.0;
-};
-
-/// margins represents the user margins for a segment.
-struct margins {
-
-    margins() = default;
-
-    explicit margins(const rj::Value::Object& val) { parse(val); };
-
-    void parse(const rj::Value::Object& val) {
-
-        rju::_getIfExists(val, enabled, "enabled");
-        rju::_getIfExists(val, net, "net");
-
-        rj::Value avlVal(rj::kObjectType);
-        rju::_getIfExists(val, avlVal, "available", rju::_RJValueType::Object);
-        rj::Value usedVal(rj::kObjectType);
-        rju::_getIfExists(val, usedVal, "utilised", rju::_RJValueType::Object);
-
-        available.parse(avlVal.GetObject());
-        used.parse(usedVal.GetObject());
-    };
-
-    bool enabled = false;
-    double net = 0.0;
-    availableMargins available;
-    usedMargins used;
-};
-
-/// allMargins contains both equity and commodity margins.
-struct allMargins {
-
-    allMargins() = default;
-
-    explicit allMargins(const rj::Value::Object& val) { parse(val); };
-
-    void parse(const rj::Value::Object& val) {
-
-        rj::Value eqVal(rj::kObjectType);
-        rju::_getIfExists(val, eqVal, "equity", rju::_RJValueType::Object);
-        equity.parse(eqVal.GetObject());
-
-        rj::Value cmVal(rj::kObjectType);
-        rju::_getIfExists(val, cmVal, "commodity", rju::_RJValueType::Object);
-        commodity.parse(cmVal.GetObject());
-    };
-
-    margins equity;
-    margins commodity;
-};
-
 /// ohlc strcut
 struct ohlc {
