+++ include/kitepp/responses/responses.hpp
@@ -5,3 +5,4 @@
 #include "mf.hpp"
 #include "order.hpp"
 #include "portfolio.hpp"
+#include "user.hpp"
